Add files via upload
Here is a rough hacked together version of the code! It ain't good in the slightest and I will make some improvements later down the line, but for now I am happy that it works! On my PC, after about 5 hours, it will give me a seed that when doing random.randint(1,10) ten times, it will give me all the numbers 1-10 in ascending order! Tbh, idk how to make it faster than this while still using Python, Python is rather slowww....