Code improvement + slight speed increase
I simplified the code to make it not so messy! Simplifying the code didn't seem to have any significant effects on the speed (although it looks prettier now), but for some reason putting everything into a function sped up the code by a little bit? I read somewhere that the reason for this is because things inside a function are done locally, whereas things outside the function are global? I'm not entirely sure what it means, but I'm happy about the speed increase either way :)