Code overhaul
I remade the entire code to try squeeze as much performance from it as I possibly could. I found that instead of using a loop to generate numbers from 1 to n, I could use a list to create a custom order and the time impact from that would be negligible. I also remade the brute force loop to squeeze out as much time as I possibly could from it, including using a function from the itertools module that once again slightly improved speed. I then remembered hearing about multiprocessing, so I decided to use that to once again further increase the speed - leading to the greatest speed improvement thus far! I can think of minor changes that could be made to better this program, but I don't think it is worth it to add them as it leaves the scope of what I wanted to do with this small project :)